pyproject.toml
[project]
name = "pubmed-fetcher"
version = "0.1.0"
description = ""
authors = [
    {name = "Karthik",email = "karthikvallabhaneni3712@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[tool.poetry.scripts]
get-papers-list = "pubmed_fetcher.pubmed:main"


[tool.poetry]
packages = [{include = "pubmed_fetcher", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
